[python/zcl] Move generated parts of Python CHIP controller
Move both C++ and Python ZCL pieces of Python CHIP
controller into a single directory.
