+++ src/controller/python/BUILD.gn
@@ -36,7 +36,6 @@ shared_library("ChipDeviceCtrl") {
   include_dirs = [ "." ]
 
   sources = [
-    "ChipDeviceController-ClusterCommands.cpp",
     "ChipDeviceController-ScriptBinding.cpp",
     "ChipDeviceController-ScriptDeviceAddressUpdateDelegate.cpp",
     "ChipDeviceController-ScriptDeviceAddressUpdateDelegate.h",
@@ -44,6 +43,7 @@ shared_library("ChipDeviceCtrl") {
     "ChipDeviceController-ScriptDevicePairingDelegate.h",
     "ChipDeviceController-StorageDelegate.cpp",
     "ChipDeviceController-StorageDelegate.h",
+    "chip/clusters/CHIPClusters.cpp",
     "chip/internal/ChipThreadWork.cpp",
     "chip/internal/ChipThreadWork.h",
     "chip/internal/CommissionerImpl.cpp",
@@ -89,7 +89,6 @@ pw_python_action("python") {
         "chip/ChipBleBase.py",
         "chip/ChipBleUtility.py",
         "chip/ChipBluezMgr.py",
-        "chip/ChipCluster.py",
         "chip/ChipCoreBluetoothMgr.py",
         "chip/ChipDeviceCtrl.py",
         "chip/ChipStack.py",
@@ -101,6 +100,7 @@ pw_python_action("python") {
         "chip/ble/library_handle.py",
         "chip/ble/scan_devices.py",
         "chip/ble/types.py",
+        "chip/clusters/CHIPClusters.py",
         "chip/configuration/__init__.py",
         "chip/exceptions/__init__.py",
         "chip/internal/__init__.py",
